// packages/blog-starter-kit/themes/enterprise/generated/graphql.ts
export type Maybe<T> = T | null;
// ...
export type Scalars = {
	ID: { input: string; output: string };
	String: { input: string; output: string };
	Boolean: { input: boolean; output: boolean };
	Int: { input: number; output: number };
	Float: { input: number; output: number };
	DateTime: { input: string; output: string };
	ObjectId: { input: string; output: string };
	TimeZone: { input: any; output: any };
	URL: { input: any; output: any };
};
// ...
export type Badge = Node & {
	__typename?: 'Badge';
	/** The date the badge was earned. */
	dateAssigned?: Maybe<Scalars['DateTime']['output']>;
	/** The description of the badge. */
	description?: Maybe<Scalars['String']['output']>;
	/** The ID of the badge. */
	id: Scalars['ID']['output'];
	/** The image of the badge. */
	image: Scalars['String']['output'];
	/** Link to badge page on Hashnode. */
	infoURL?: Maybe<Scalars['String']['output']>;
	/** The name of the badge. */
	name: Scalars['String']['output'];
	/** A flag to determine if the badge is hidden. */
	suppressed?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type Content = {
	__typename?: 'Content';
	/** The HTML version of the content. */
	html: Scalars['String']['output'];
	/** The Markdown version of the content. */
	markdown: Scalars['String']['output'];
	/** The text version from sanitized html content. HTML tags are stripped and only text is returned. */
	text: Scalars['String']['output'];
};
// ...
export type IUser = {
	/** Whether or not the user is an ambassador. */
	ambassador: Scalars['Boolean']['output'];
	/** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
	availableFor?: Maybe<Scalars['String']['output']>;
	/** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
	badges: Array<Badge>;
	/** The bio of the user. Visible in about me section of the user's profile. */
	bio?: Maybe<Content>;
	/** The date the user joined Hashnode. */
	dateJoined?: Maybe<Scalars['DateTime']['output']>;
	/** Whether or not the user is deactivated. */
	deactivated: Scalars['Boolean']['output'];
	/** The users who are following this user */
	followers: UserConnection;
	/** The number of users that follow the requested user. Visible in the user's profile. */
	followersCount: Scalars['Int']['output'];
	/** The number of users that this user is following. Visible in the user's profile. */
	followingsCount: Scalars['Int']['output'];
	/** The users which this user is following */
	follows: UserConnection;
	/** The ID of the user. It can be used to identify the user. */
	id: Scalars['ID']['output'];
	/** The location of the user. */
	location?: Maybe<Scalars['String']['output']>;
	/** The name of the user. */
	name: Scalars['String']['output'];
	/** Returns the list of posts the user has published. */
	posts: UserPostConnection;
	/** The URL to the profile picture of the user. */
	profilePicture?: Maybe<Scalars['String']['output']>;
	/** Publications associated with the user. Includes personal and team publications. */
	publications: UserPublicationsConnection;
	/** The social media links of the user. Shown on the user's profile. */
	socialMediaLinks?: Maybe<SocialMediaLinks>;
	/** The tagline of the user. Shown on the user's profile below the name. */
	tagline?: Maybe<Scalars['String']['output']>;
	/** Returns a list of tags that the user follows. */
	tagsFollowing: Array<Tag>;
	/** Returns list of tags from user's expertise. Shown on the user's profile. */
	techStack: UserTagsConnection;
	/** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
	username: Scalars['String']['output'];
};
// ...
export type Node = {
	/** The ID of the node. */
	id: Scalars['ID']['output'];
};
// ...
export type PostCoverImage = {
	__typename?: 'PostCoverImage';
	/** Provides attribution information for the cover image, if available. */
	attribution?: Maybe<Scalars['String']['output']>;
	/** True if the image attribution should be hidden. */
	isAttributionHidden: Scalars['Boolean']['output'];
	/** Indicates whether the cover image is in portrait orientation. */
	isPortrait: Scalars['Boolean']['output'];
	/** The name of the photographer who captured the cover image. */
	photographer?: Maybe<Scalars['String']['output']>;
	/** The URL of the cover image. */
	url: Scalars['String']['output'];
};
// ...
export type SocialMediaLinks = {
	__typename?: 'SocialMediaLinks';
	/** The user's Bluesky profile. */
	bluesky?: Maybe<Scalars['String']['output']>;
	/** The user's Facebook profile. */
	facebook?: Maybe<Scalars['String']['output']>;
	/** The user's GitHub profile. */
	github?: Maybe<Scalars['String']['output']>;
	/** The user's Instagram profile. */
	instagram?: Maybe<Scalars['String']['output']>;
	/** The user's LinkedIn profile. */
	linkedin?: Maybe<Scalars['String']['output']>;
	/** The user's StackOverflow profile. */
	stackoverflow?: Maybe<Scalars['String']['output']>;
	/** The user's Twitter profile. */
	twitter?: Maybe<Scalars['String']['output']>;
	/** The user's website. */
	website?: Maybe<Scalars['String']['output']>;
	/** The user's YouTube profile. */
	youtube?: Maybe<Scalars['String']['output']>;
};
// ...
export type Tag = ITag &
	Node & {
		__typename?: 'Tag';
		/** Total number of users following this tag. */
		followersCount: Scalars['Int']['output'];
		/** The ID of the tag. */
		id: Scalars['ID']['output'];
		/** Information about the tag. Contains markdown html and text version of the tag's info. */
		info?: Maybe<Content>;
		/** The logo of the tag. Shown in tag page. */
		logo?: Maybe<Scalars['String']['output']>;
		/** The name of the tag. Shown in tag page. */
		name: Scalars['String']['output'];
		/** Paginated list of posts published under this tag */
		posts: FeedPostConnection;
		/** Alltime usage count of this tag in posts. */
		postsCount: Scalars['Int']['output'];
		/** The slug of the tag. Used to access tags feed.  Example https://hashnode.com/n/graphql */
		slug: Scalars['String']['output'];
		/** The tagline of the tag. */
		tagline?: Maybe<Scalars['String']['output']>;
	};
// ...
export type User = IUser &
	Node & {
		__typename?: 'User';
		/**
		 * Whether or not the user is an ambassador.
		 * @deprecated Ambassadors program no longer active. Will be removed after 02/01/2024
		 */
		ambassador: Scalars['Boolean']['output'];
		/** The availability of the user based on tech stack and interests. Shown on the "I am available for" section in user's profile. */
		availableFor?: Maybe<Scalars['String']['output']>;
		/** Returns a list of badges that the user has earned. Shown on blogs /badges page. Example - https://iamshadmirza.com/badges */
		badges: Array<Badge>;
		/** The bio of the user. Visible in about me section of the user's profile. */
		bio?: Maybe<Content>;
		/**
		 * The bio of the user. Visible in about me section of the user's profile.
		 * @deprecated Will be removed on 26/10/2023. Use bio instead of bioV2
		 */
		bioV2?: Maybe<Content>;
		/** The date the user joined Hashnode. */
		dateJoined?: Maybe<Scalars['DateTime']['output']>;
		/** Whether or not the user is deactivated. */
		deactivated: Scalars['Boolean']['output'];
		/** The users who are following this user */
		followers: UserConnection;
		/** The number of users that follow the requested user. Visible in the user's profile. */
		followersCount: Scalars['Int']['output'];
		/**
		 * Whether or not the authenticated user follows this user.
		 * Returns false if the authenticated user this user.
		 */
		following: Scalars['Boolean']['output'];
		/** The number of users that this user is following. Visible in the user's profile. */
		followingsCount: Scalars['Int']['output'];
		/** The users which this user is following */
		follows: UserConnection;
		/**
		 * Whether or not this user follows the authenticated user.
		 * Returns false if the authenticated user this user.
		 */
		followsBack: Scalars['Boolean']['output'];
		/** The ID of the user. It can be used to identify the user. */
		id: Scalars['ID']['output'];
		/** The location of the user. */
		location?: Maybe<Scalars['String']['output']>;
		/** The name of the user. */
		name: Scalars['String']['output'];
		/** Returns the list of posts the user has published. */
		posts: UserPostConnection;
		/** The URL to the profile picture of the user. */
		profilePicture?: Maybe<Scalars['String']['output']>;
		/** Publications associated with the user. Includes personal and team publications. */
		publications: UserPublicationsConnection;
		/** The social media links of the user. Shown on the user's profile. */
		socialMediaLinks?: Maybe<SocialMediaLinks>;
		/** The tagline of the user. Shown on the user's profile below the name. */
		tagline?: Maybe<Scalars['String']['output']>;
		/** Returns a list of tags that the user follows. */
		tagsFollowing: Array<Tag>;
		/** Returns list of tags from user's expertise. Shown on the user's profile. */
		techStack: UserTagsConnection;
		/** The username of the user. It is unique and tied with user's profile URL. Example - https://hashnode.com/@username */
		username: Scalars['String']['output'];
	};
// ...
export type UserConnection = PageConnection & {
	__typename?: 'UserConnection';
	/** A list of users */
	nodes: Array<User>;
	/** Information for page based pagination in users connection. */
	pageInfo: OffsetPageInfo;
	/** The total number of documents in the connection. */
	totalDocuments: Scalars['Int']['output'];
};
// ...
export type UserPostConnection = PageConnection & {
	__typename?: 'UserPostConnection';
	/** A list of edges containing Post information */
	edges: Array<UserPostEdge>;
	/** A list of posts */
	nodes: Array<Post>;
	/** Information for page based pagination in Post connection. */
	pageInfo: OffsetPageInfo;
	/** The total number of documents in the connection. */
	totalDocuments: Scalars['Int']['output'];
};
// ...
export type UserPublicationsConnection = Connection & {
	__typename?: 'UserPublicationsConnection';
	/** A list of edges of publications connection. */
	edges: Array<UserPublicationsEdge>;
	/** Information to aid in pagination. */
	pageInfo: PageInfo;
	/** The total amount of publications taking into account the filter. */
	totalDocuments: Scalars['Int']['output'];
};
// ...
export type UserTagsConnection = PageConnection & {
	__typename?: 'UserTagsConnection';
	/** A list of tags */
	nodes: Array<Tag>;
	/** Information for page based pagination in users connection. */
	pageInfo: OffsetPageInfo;
	/** The total number of documents in the connection. */
	totalDocuments: Scalars['Int']['output'];
};
// ...
export type PostFragment = {
	__typename?: 'Post';
	id: string;
	title: string;
	url: string;
	publishedAt: string;
	slug: string;
	brief: string;
	author: { __typename?: 'User'; name: string; profilePicture?: string | null };
	coverImage?: { __typename?: 'PostCoverImage'; url: string } | null;
};